Monthly reports

+++ src/components/expense/services/expense.service.ts
@@ -5,12 +5,13 @@ import {
   forwardRef,
   InternalServerErrorException,
   Logger,
+  UnauthorizedException,
 } from '@nestjs/common';
 
 import { Expense } from '../schemas/expense.schema';
 import { CreateExpenseDto } from '../dtos/expense.create.dto';
 import { UpdateExpenseDto } from '../dtos/expense.update.dto';
-import { Model, Types } from 'mongoose';
+import mongoose, { Model, Types } from 'mongoose';
 import { InjectModel } from '@nestjs/mongoose';
 import {
   UserRepository,
@@ -22,6 +23,8 @@ import {
 } from 'src/components/category/repos/category.repository';
 import { Attachment } from 'src/components/attachment/schemas/attachment.schema';
 import { AttachmentService } from 'src/components/attachment/services/attachment.service';
+import { CreateAttachmentDto } from 'src/components/attachment/dtos/attachment.create.dto';
+import * as fs from 'fs';
 
 @Injectable()
 export class ExpenseService {
@@ -37,18 +40,19 @@ export class ExpenseService {
   async create(
     createExpenseDto: CreateExpenseDto,
     userId: string,
+    newAttachment: CreateAttachmentDto | null,
   ): Promise<Expense> {
     const session = await this.expenseModel.db.startSession(); //we start a session here of dif queries
     session.startTransaction(); //incase of an error, all queries, won't take effect
 
     try {
-      const categoryName = createExpenseDto.category;
+      const categoryName = createExpenseDto.categoryId;
       const categId = await this.categoryRepository.findOneByName(
         categoryName,
         userId,
         session,
       );
-      delete createExpenseDto.category;
+      delete createExpenseDto.categoryId;
 
       const dateInstance = new Date(createExpenseDto.date);
       delete createExpenseDto.date;
@@ -57,19 +61,43 @@ export class ExpenseService {
         date: dateInstance,
         userId: userId,
         categoryId: categId,
+        attachment: null, // we start off with a null attachment
       };
+
       const createdExpense = new this.expenseModel(expense);
+
+      //if there is an attachment
+      if (newAttachment) {
+        console.log('newAttachment: ', newAttachment);
+        newAttachment.expenseId = createdExpense._id; //we set the expenseId to the attachment
+
+        //we create the attachment
+        const createdAttachment = await this.attachmentService.create(
+          newAttachment,
+        );
+
+        //we update the expenseId of the newly created attachment
+        await this.attachmentService.updateExpenseId(
+          createdAttachment.id,
+          createdExpense.id,
+        );
+
+        // we update the attachment id of the expense
+        createdExpense.attachment = createdAttachment._id; // add the attachment id to the expense
+        await createdExpense.save({ session });
+      }
 
       await this.userRepository.addExpenseToUser(
         userId,
         new Types.ObjectId(createdExpense.id),
         session,
-      ); // Method which adds the category ID to the user who created it
+      ); // Method which adds the expense ID to the user who created it
 
       await this.categoryRepository.addExpenseToCategory(
         expense.categoryId,
         new Types.ObjectId(createdExpense.id),
         createdExpense.amount,
+        createdExpense.date,
         session,
       );
 
@@ -100,20 +128,228 @@ export class ExpenseService {
     return expense;
   }
 
-  async update(id: string, expenseDto: UpdateExpenseDto): Promise<Expense> {
-    let updatedExpense: any = { ...expenseDto };
-    const expense = await this.expenseModel
-      .findByIdAndUpdate(id, updatedExpense, {
-        new: true,
-        runValidators: true,
-      })
-      .exec();
+  async getOverview(
+    month: number,
+    year: number,
+    userId: string,
+  ): Promise<Expense[]> {
+    Logger.log('dto month zzz: ', month);
+    Logger.log('dto month zzz: ', year);
 
-    if (!expense) {
-      throw new NotFoundException(`Expense with ID '${id}' not found`);
+    const startDate = new Date(year, month - 1); // month starts at zero, so dec is 0
+    const endDate =
+      month == 12 ? new Date(+year + 1, 0) : new Date(year, month % 12); // month % 12 will give 0 for December
+
+    Logger.log('startDate: ', startDate);
+    Logger.log('endDate: ', endDate);
+
+    const expenses = await this.expenseModel.aggregate([
+      {
+        $match: {
+          userId: userId,
+          date: { $gte: startDate, $lt: endDate },
+        },
+      },
+      {
+        $lookup: {
+          from: 'categories',
+          let: { categoryId: { $toObjectId: '$categoryId' } },
+          pipeline: [{ $match: { $expr: { $eq: ['$_id', '$$categoryId'] } } }],
+          as: 'categoryData',
+        },
+      },
+      {
+        $unwind: '$categoryData',
+      },
+      {
+        $project: {
+          _id: 0,
+          date: { $dateToString: { format: '%d-%m-%Y', date: '$date' } },
+          concept: '$concept',
+          category: '$categoryData.name',
+          totalAmount: '$amount',
+        },
+      },
+    ]);
+
+    return expenses;
+  }
+
+  async getCategories(
+    month: number,
+    year: number,
+    userId: string,
+  ): Promise<Expense[]> {
+    const startDate = new Date(year, month - 1); // month starts at zero, so dec is 0
+    const endDate =
+      month == 12 ? new Date(+year + 1, 0) : new Date(year, month % 12); // month % 12 will give 0 for December
+
+    const expenses = await this.expenseModel.aggregate([
+      {
+        $match: {
+          userId: userId,
+          date: { $gte: startDate, $lt: endDate },
+        },
+      },
+      {
+        $lookup: {
+          from: 'categories',
+          let: { categoryId: { $toObjectId: '$categoryId' } },
+          pipeline: [{ $match: { $expr: { $eq: ['$_id', '$$categoryId'] } } }],
+          as: 'categoryData',
+        },
+      },
+      {
+        $unwind: '$categoryData',
+      },
+      {
+        $group: {
+          _id: '$categoryData.name',
+          total: { $sum: '$amount' },
+          min: { $first: '$categoryData.minValue' },
+          max: { $first: '$categoryData.maxValue' },
+        },
+      },
+      {
+        $project: {
+          _id: 0,
+          category: '$_id',
+          totalAmount: '$total',
+          min: '$min',
+          max: '$max',
+          status: {
+            $cond: {
+              if: { $gt: ['$total', '$max'] },
+              then: 'Exceeding',
+              else: {
+                $cond: {
+                  if: { $lt: ['$total', '$min'] },
+                  then: 'Saving',
+                  else: 'Within Limit',
+                },
+              },
+            },
+          },
+        },
+      },
+    ]);
+
+    return expenses;
+  }
+
+  async update(
+    id: string,
+    expenseDto: UpdateExpenseDto,
+    userId: string,
+    newAttachment: CreateAttachmentDto | null,
+  ): Promise<Expense> {
+    const session = await this.expenseModel.db.startSession();
+    session.startTransaction();
+    // check if the category has been changed
+
+    try {
+      const existingExpense = await this.expenseModel.findById(id).exec();
+      if (!existingExpense) {
+        throw new NotFoundException(`Expense with ID '${id}' not found`);
+      }
+      // Check if the user has permissions to update the expense
+      if (existingExpense.userId !== userId) {
+        throw new UnauthorizedException(
+          'You do not have permission to update this expense',
+        );
+      }
+
+      // const { attachment } = existingExpense; // we destruct the attachment property from the expense object
+
+      //if there is an attachment
+      if (newAttachment) {
+        console.log('newAttachment: ', newAttachment);
+        newAttachment.expenseId = existingExpense._id; //we set the expenseId to the attachment
+
+        //we create the attachment
+        const createdAttachment = await this.attachmentService.create(
+          newAttachment,
+        );
+
+        //we update the expenseId of the newly created attachment
+        await this.attachmentService.updateExpenseId(
+          createdAttachment.id,
+          existingExpense.id,
+        );
+
+        // we update the attachment id of the expense
+        existingExpense.attachment = createdAttachment._id; // add the attachment id to the expense
+      }
+      //we have the categoryName but we need the category Id
+      const categoryName = expenseDto.categoryId;
+      const categId = await this.categoryRepository.findOneByName(
+        categoryName,
+        existingExpense.userId,
+        session,
+      );
+      delete expenseDto.categoryId;
+
+      const dateInstance = new Date(expenseDto.date);
+      delete expenseDto.date;
+      const updatedExpense = {
+        ...expenseDto,
+        date: dateInstance,
+        userId: existingExpense.userId,
+        categoryId: categId,
+        attachment: newAttachment
+          ? existingExpense.attachment
+          : expenseDto.attachment,
+      };
+
+      const expense = await this.expenseModel
+        .findByIdAndUpdate(id, updatedExpense, {
+          new: true,
+          runValidators: true,
+        })
+        .exec();
+
+      if (!expense) {
+        throw new NotFoundException(`Expense with ID '${id}' not found`);
+      }
+
+      if (existingExpense.categoryId !== expense.categoryId) {
+        // remove the expense from the old category and deduct from current value if its in the same month/year of today
+        await this.categoryRepository.deleteExpenseFromCategory(
+          existingExpense.categoryId,
+          new Types.ObjectId(id),
+          existingExpense.amount,
+          existingExpense.date,
+          session,
+        );
+
+        // add the expense to the new category and add to current value if its in the same month/year of today
+        await this.categoryRepository.addExpenseToCategory(
+          expense.categoryId,
+          new Types.ObjectId(id),
+          expense.amount,
+          expense.date,
+          session,
+        );
+      } else {
+        await this.categoryRepository.updateExpenseInCategory(
+          expense.categoryId,
+          new Types.ObjectId(id),
+          existingExpense.amount,
+          expense.amount,
+          expense.date,
+          session,
+        );
+      }
+
+      await session.commitTransaction();
+      return expense;
+    } catch (error) {
+      // if anything goes wrong, undo any changes made in the database
+      await session.abortTransaction();
+      throw error;
+    } finally {
+      session.endSession();
     }
-
-    return expense;
   }
 
   async updateToUnCategorized(
@@ -130,6 +366,7 @@ export class ExpenseService {
   }
 
   async delete(id: string): Promise<Expense> {
+    Logger.log('id: ', id);
     const session = await this.expenseModel.db.startSession(); //we start a session here of dif queries
     session.startTransaction(); //incase of an error, all queries, won't take effect
 
@@ -141,6 +378,7 @@ export class ExpenseService {
       if (!deletedExpense) {
         throw new NotFoundException(`Expense with ID '${id}' not found`);
       }
+      Logger.log('deletedExpense: ', deletedExpense);
 
       await this.userRepository.deleteExpenseFromUser(
         deletedExpense.userId,
@@ -149,15 +387,21 @@ export class ExpenseService {
       );
 
       await this.categoryRepository.deleteExpenseFromCategory(
-        id,
         deletedExpense.categoryId,
+        new Types.ObjectId(id),
+        deletedExpense.amount,
+        deletedExpense.date,
         session,
       );
 
-      await this.attachmentService.deleteAttachmentOfExpense(
-        deletedExpense.attachment.id,
-        session,
-      );
+      //if there's an attachment, we delete it too
+      if (deletedExpense.attachment) {
+        console.log('deletedExpense.attachment: ', deletedExpense.attachment);
+        await this.attachmentService.deleteAttachmentOfExpense(
+          deletedExpense.attachment.toString(),
+          session,
+        );
+      }
 
       await session.commitTransaction();
       return deletedExpense;
@@ -192,47 +436,43 @@ export class ExpenseService {
     }
 
     expense.attachment = null;
+    console.log('removeAttachmentFromExpense: ', expense);
+
     await expense.save();
   }
 
-  async addAttachmentToExpense(
-    expenseId: string,
-    attachment: Attachment,
+  async deleteAttachmentFile(
+    storageLocation: string,
     session: any,
   ): Promise<void> {
-    const expense = await this.expenseModel
-      .findById(expenseId)
-      .session(session);
-
-    if (!expense) {
-      throw new NotFoundException(`Expense with id '${expenseId}' not found`);
+    try {
+      fs.unlinkSync(storageLocation); // Delete the file in storage location
+    } catch (error) {
+      // Handle error appropriately
+      console.error(error);
     }
-
-    expense.attachment = attachment;
-    await expense.save();
-  }
-
-  async findAttachmentOfExpense(expenseId: string): Promise<Attachment> {
-    const expense2 = await this.expenseModel.findById(expenseId).exec();
-    if (!expense2) {
-      throw new NotFoundException(`Expense with ID '${expenseId}' not found`);
-    }
-    expense2.populate('attachment');
-
-    const expense = await this.expenseModel
-      .findById(expenseId)
-      .populate('attachment')
-      .exec();
-
-    if (!expense) {
-      throw new NotFoundException(`Expense with id '${expenseId}' not found`);
-    }
-
-    if (!expense.attachment) {
-      throw new NotFoundException(
-        `Expense with id ${expenseId} doesn't have an attachment`,
-      );
-    }
-    return expense.attachment;
   }
 }
+//   async findAttachmentOfExpense(expenseId: string): Promise<Attachment> {
+//   const expense2 = await this.expenseModel.findById(expenseId).exec();
+//   if (!expense2) {
+//     throw new NotFoundException(`Expense with ID '${expenseId}' not found`);
+//   }
+//   expense2.populate('attachment');
+
+//   const expense = await this.expenseModel
+//     .findById(expenseId)
+//     .populate('attachment')
+//     .exec();
+
+//   if (!expense) {
+//     throw new NotFoundException(`Expense with id '${expenseId}' not found`);
+//   }
+
+//   if (!expense.attachment) {
+//     throw new NotFoundException(
+//       `Expense with id ${expenseId} doesn't have an attachment`,
+//     );
+//   }
+//   return expense.attachment;
+// }
